cos_mama_project_backend-main/src/test/java/com/cosmama/backend/UserControllerTest.java
package com.cosmama.backend;

import com.cosmama.backend.api.controller.UserController;
import com.cosmama.backend.api.model.User;
import com.cosmama.backend.api.service.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.util.Optional;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;

@WebMvcTest(UserController.class)
class UserControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @BeforeEach
    void setUp() {
        User user = new User(1L, "Test", "User", "test@example.com");
        Mockito.when(userService.getUser(1L)).thenReturn(Optional.of(user));
        Mockito.when(userService.getUser(99L)).thenReturn(Optional.empty());
    }

    @Test
    void getUser_ReturnsUser_IfExists() throws Exception {
        mockMvc.perform(get("/user?id=1")
                        .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.id").value(1))
                .andExpect(jsonPath("$.firstName").value("Test"))
                .andExpect(jsonPath("$.lastName").value("User"))
                .andExpect(jsonPath("$.email").value("test@example.com"));
    }

    @Test
    void getUser_ReturnsNotFound_IfNotExists() throws Exception {
        mockMvc.perform(get("/user?id=99")
                        .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$").doesNotExist());
    }
}